renamed native_game module
